Changed health and damage values to int, corrected Nut class

+++ src/main/java/flowerforce/model/entities/AbstractPlant.java
@@ -18,7 +18,7 @@ public abstract class AbstractPlant extends AbstractLivingEntity implements Plan
      * @param health the starting health of the entity
      * @param plantType the type of the plant
      */
-    public AbstractPlant(final Point2D pos, final Timer timer, final double health, final Plants plantType) {
+    public AbstractPlant(final Point2D pos, final Timer timer, final int health, final Plants plantType) {
         super(pos, timer, health);
         this.plantType = plantType;
     }
