+++ src/main/java/flowerforce/model/entities/SunflowerImpl.java
@@ -16,7 +16,7 @@ public class SunflowerImpl extends AbstractPlant implements Sunflower {
      * @param pos the initial position to place the entity in
      * @param plantType the type of the plant
      */
-    public SunflowerImpl(final Point2D pos, final Timer sunProductionTimer, final double health, final Plants plantType) {
+    public SunflowerImpl(final Point2D pos, final Timer sunProductionTimer, final int health, final Plants plantType) {
         super(pos, sunProductionTimer, health, plantType);
     }
